+++ StripeCore/StripeCore/Source/API Bindings/STPAPIClient.swift
@@ -264,7 +264,7 @@ import UIKit
 
 }
 
-private let APIVersion = "2020-08-27"
+private let APIVersion = "2026-07-29.dahlia"
 private let APIBaseURL = "https://api.stripe.com/v1"
 
 // MARK: Modern bindings